When changing reaction scenario, adapt the parameters of the previously selected scenario, so that the behavior of the reaction is "more or less" the same
git-svn-id: http://hmisvn.ewi.utwente.nl/ExBio/projects/inat@71 d52acf90-180f-11e0-a790-6b35089b84c1

+++ src/inat/model/Scenario.java
@@ -44,7 +44,7 @@ public class Scenario {
 				double par1 = parameters.get(SCENARIO_PARAMETER_K2_KM),
 					   Stot = parameters.get(SCENARIO_PARAMETER_STOT),
 					   S = Stot - r2Level,
-					   E = r1Level;
+					   E = (double)r1Level;
 				double rate = par1 * E * S;
 				if (rate != 0) {
 					return Math.max(1, (int)Math.round(1 / rate));
